+++ packages/subscription-controller/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add product-scoped entitlements to `SubscriptionController` state and export Money Account Plus entitlement selectors ([#10017](https://github.com/MetaMask/core/pull/10017))
+- Add product-scoped entitlements to `SubscriptionController` state and export type-safe product entitlement selectors for Money Account Plus and Shield ([#10017](https://github.com/MetaMask/core/pull/10017))
 
 ## [8.0.1]